floatchatserver/requirements.txt
django>=4.2
python-dotenv
requests
pandas
numpy
faiss-cpu
sentence-transformers
google-generativeai
